Pull mass matrix and preconditioner into classes.
